# swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/health": {
      "get": {
        "operationId": "AppController_getHealth",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешный вход",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Неверные учетные данные"
          }
        },
        "summary": "Вход в систему",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Токены обновлены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Неверный refresh токен"
          }
        },
        "summary": "Обновление токенов",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Выход выполнен успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Необходима авторизация"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Выход из системы",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Данные пользователя",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserDto"
                }
              }
            }
          },
          "401": {
            "description": "Необходима авторизация"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Получить текущего пользователя",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/users/owner": {
      "post": {
        "operationId": "UsersController_createOwner",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOwnerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Собственник создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации"
          },
          "401": {
            "description": "Необходима авторизация"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Email уже существует"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Создать собственника (только админ)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/manager": {
      "post": {
        "operationId": "UsersController_createManager",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateManagerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Менеджер создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации"
          },
          "401": {
            "description": "Необходима авторизация"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Email уже существует"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Создать менеджера (только админ)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/managers": {
      "get": {
        "operationId": "UsersController_getManagers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Список менеджеров",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Необходима авторизация"
          },
          "403": {
            "description": "Недостаточно прав"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Список менеджеров (только админ)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}": {
      "delete": {
        "operationId": "UsersController_deleteManager",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Менеджер удалён"
          },
          "401": {
            "description": "Необходима авторизация"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "404": {
            "description": "Менеджер не найден"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Удалить менеджера (только админ)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/manager-properties": {
      "post": {
        "operationId": "ManagerPropertiesController_assign",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignPropertyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Назначение создано",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagerPropertyResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации"
          },
          "404": {
            "description": "Менеджер или объект не найден"
          },
          "409": {
            "description": "Объект уже назначен"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Назначить объект на менеджера",
        "tags": [
          "Manager Properties"
        ]
      },
      "get": {
        "operationId": "ManagerPropertiesController_findAll",
        "parameters": [
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "propertyId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список назначений",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ManagerPropertyResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Список назначений (фильтр по userId/propertyId)",
        "tags": [
          "Manager Properties"
        ]
      }
    },
    "/api/manager-properties/{id}": {
      "delete": {
        "operationId": "ManagerPropertiesController_unassign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Назначение удалено"
          },
          "404": {
            "description": "Назначение не найдено"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Снять назначение объекта",
        "tags": [
          "Manager Properties"
        ]
      }
    }
  },
  "info": {
    "title": "Balivi API",
    "description": "API для управления арендой недвижимости",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Введите JWT токен",
        "in": "header"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email пользователя",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "Пароль пользователя",
            "example": "password123",
            "minLength": 6
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID пользователя",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "email": {
            "type": "string",
            "description": "Email пользователя",
            "example": "user@example.com"
          },
          "name": {
            "type": "string",
            "description": "Имя пользователя",
            "example": "Иван Иванов"
          },
          "role": {
            "type": "string",
            "description": "Роль пользователя",
            "enum": [
              "ADMIN",
              "MANAGER",
              "OWNER"
            ],
            "example": "OWNER"
          },
          "ownerId": {
            "type": "object",
            "description": "ID собственника (если роль OWNER)",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "nullable": true
          },
          "canCreateOwners": {
            "type": "boolean",
            "description": "Может создавать собственников (делегировано админом)",
            "example": false
          },
          "canCreateProperties": {
            "type": "boolean",
            "description": "Может создавать объекты (делегировано админом)",
            "example": false
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role",
          "ownerId",
          "canCreateOwners",
          "canCreateProperties"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT access токен",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "description": "JWT refresh токен",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "user": {
            "description": "Данные пользователя",
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthUserDto"
              }
            ]
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "user"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh токен",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "TokenPairResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT access токен",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "description": "JWT refresh токен",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ]
      },
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Сообщение",
            "example": "Выход выполнен успешно"
          }
        },
        "required": [
          "message"
        ]
      },
      "CreateOwnerDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email собственника",
            "example": "owner@example.com"
          },
          "password": {
            "type": "string",
            "description": "Пароль (минимум 6 символов)",
            "example": "password123",
            "minLength": 6
          },
          "name": {
            "type": "string",
            "description": "Имя собственника",
            "example": "Иван Иванов"
          },
          "phone": {
            "type": "string",
            "description": "Телефон собственника",
            "example": "+7 999 123-45-67"
          }
        },
        "required": [
          "email",
          "password",
          "name"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID пользователя",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "email": {
            "type": "string",
            "description": "Email пользователя",
            "example": "owner@example.com"
          },
          "name": {
            "type": "string",
            "description": "Имя пользователя",
            "example": "Иван Иванов"
          },
          "role": {
            "type": "string",
            "description": "Роль пользователя",
            "enum": [
              "ADMIN",
              "MANAGER",
              "OWNER"
            ],
            "example": "OWNER"
          },
          "ownerId": {
            "type": "object",
            "description": "ID собственника (если роль OWNER)",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "nullable": true
          },
          "phone": {
            "type": "object",
            "description": "Телефон собственника",
            "example": "+7 999 123-45-67"
          },
          "canCreateOwners": {
            "type": "boolean",
            "description": "Может создавать собственников",
            "example": false
          },
          "canCreateProperties": {
            "type": "boolean",
            "description": "Может создавать объекты",
            "example": false
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Дата создания",
            "example": "2026-02-23T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role",
          "ownerId",
          "canCreateOwners",
          "canCreateProperties",
          "createdAt"
        ]
      },
      "CreateManagerDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email менеджера",
            "example": "manager@example.com"
          },
          "password": {
            "type": "string",
            "description": "Пароль (минимум 6 символов)",
            "example": "password123",
            "minLength": 6
          },
          "name": {
            "type": "string",
            "description": "Имя менеджера",
            "example": "Петр Петров"
          },
          "canCreateOwners": {
            "type": "boolean",
            "description": "Может создавать собственников",
            "example": false,
            "default": false
          },
          "canCreateProperties": {
            "type": "boolean",
            "description": "Может создавать объекты",
            "example": false,
            "default": false
          }
        },
        "required": [
          "email",
          "password",
          "name"
        ]
      },
      "AssignPropertyDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "ID менеджера",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "propertyId": {
            "type": "string",
            "description": "ID объекта",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        },
        "required": [
          "userId",
          "propertyId"
        ]
      },
      "ManagerPropertyResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID назначения"
          },
          "userId": {
            "type": "string",
            "description": "ID менеджера"
          },
          "propertyId": {
            "type": "string",
            "description": "ID объекта"
          },
          "userName": {
            "type": "string",
            "description": "Имя менеджера"
          },
          "propertyTitle": {
            "type": "string",
            "description": "Название объекта"
          },
          "assignedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Дата назначения"
          }
        },
        "required": [
          "id",
          "userId",
          "propertyId",
          "userName",
          "propertyTitle",
          "assignedAt"
        ]
      }
    }
  }
}